// src/core/http/generated/models/admin-partners/partnerCouponDiscountType.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Partners
 * Filtered OpenAPI specification for Admin Partners.
 * OpenAPI spec version: 0.0.1
 */

export type PartnerCouponDiscountType = typeof PartnerCouponDiscountType[keyof typeof PartnerCouponDiscountType];


export const PartnerCouponDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;